Restructured keybindings folder
